gl-plugin: Move SerializedTlvStream into own file
We need to access the SerializedTlvStream also for the trampoline_pay
command in future commits. This commit moves the struct and all
associated commits and helper functions into their own module.
